requirements.txt
Flask==2.3.3
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
Werkzeug==2.3.7
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
requests==2.31.0
gunicorn==21.2.0
